Temp

+++ community/src/main/java/com/lawencon/community/pojo/paymentTransaction/PojoUpdatePaymentTransactionReq.java
@@ -1,5 +1,7 @@
 package com.lawencon.community.pojo.paymentTransaction;
 
+import java.math.BigDecimal;
+
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
@@ -16,7 +18,7 @@ public class PojoUpdatePaymentTransactionReq {
 	@NotBlank(message = "Description can't be empty")
 	private String desc;
 	@NotNull(message = "Price can't be empty")
-	private Long price;
+	private BigDecimal price;
 
 	private String fileId;
 	private String fileName;
@@ -35,10 +37,10 @@ public class PojoUpdatePaymentTransactionReq {
 	public void setDesc(String desc) {
 		this.desc = desc;
 	}
-	public Long getPrice() {
+	public BigDecimal getPrice() {
 		return price;
 	}
-	public void setPrice(Long price) {
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
 	public String getFileId() {
